#212 Move API apps into central package icekit.api
Move the images and pages API apps into a new single central
`icekit.api` app to simplify the app hierarchy and group
things more sensibly in the codebase.
