+++ icekit/api/__init__.py
@@ -1,0 +1,1 @@
+default_app_config = '%s.apps.APIConfig' % __name__
