+++ icekit/api/apps.py
@@ -1,7 +1,7 @@
 from django.apps import AppConfig
 
 
-class PagesAPIConfig(AppConfig):
+class APIConfig(AppConfig):
     name = '.'.join(__name__.split('.')[:-1])  # Package with `apps` module
     label = '_'.join(__name__.split('.')[:-1])
-    verbose_name = 'PagesAPI'
+    verbose_name = 'ICEkitAPI'
