+++ icekit/api/images/views.py
