--- icekit/images_api/__init__.py
@@ -1,1 +1,0 @@
-default_app_config = '%s.apps.ImagesAPIConfig' % __name__
