--- icekit/images_api/apps.py
@@ -1,7 +1,0 @@
-from django.apps import AppConfig
-
-
-class ImagesAPIConfig(AppConfig):
-    name = '.'.join(__name__.split('.')[:-1])  # Package with `apps` module
-    label = '_'.join(__name__.split('.')[:-1])
-    verbose_name = 'ImagesAPI'
