+++ icekit/project/urls.py
@@ -41,10 +41,7 @@ urlpatterns = patterns(
 
     # Installed apps.
     url(r'^forms/', include('forms_builder.forms.urls')),
-    # APIs
-    # TODO This URL routing technique doesn't show all endpoints in root /api/
-    url(r'^api/', include('icekit.pages_api.urls')),
-    url(r'^api/', include('icekit.images_api.urls')),
+    url(r'^api/', include('icekit.api.urls')),
 
     # Get admin URLs prefix from settings.
     # Handle admin and front-end authentication separately.
